+++ specifyweb/frontend/js_src/lib/components/Router/Routes.tsx
@@ -37,14 +37,14 @@ export const routes: RA<EnhancedRoute> = [
         index: true,
         title: schemaText.databaseSchema(),
         element: () =>
-          import('../Toolbar/Schema').then(
+          import('../Toolbar/DataModel').then(
             ({ DataModelTables }) => DataModelTables
           ),
       },
       {
         path: ':tableName',
         element: () =>
-          import('../Toolbar/Schema').then(
+          import('../Toolbar/DataModel').then(
             ({ DataModelRedirect }) => DataModelRedirect
           ),
       },
